requirements.txt
Django==3.1
djangorestframework==3.12
psycopg2-binary==2.8.6
Pillow==7.1
python-decouple
django-filter
drf-yasg
lxml
BeautifulSoup4
requests